requirements.txt
Django==2.1.3
django-registration-redux==2.0
mccabe==0.6.1
pytz==2018.7
gunicorn==19.9
psycopg2==2.7
pillow==5.3.0